Add non-arkworks impl of traits for fp
