+++ benches/sqrt.rs
@@ -1,4 +1,4 @@
-use crate::Fq;
+use decaf377::Fq;
 use ark_ff::{Field, PrimeField, Zero};
 use criterion::{criterion_group, criterion_main, BenchmarkId, Criterion};
 use decaf377::{SqrtRatioZeta, ZETA};
