+++ src/fields/fp.rs
@@ -5,6 +5,7 @@ use cfg_if::cfg_if;
 
 #[cfg(feature = "arkworks")]
 pub mod arkworks;
+mod ops;
 pub mod u32;
 pub mod u64;
